// src/Rules/Rules.types.ts
export interface RuleProps {
  score: number;
  description: string;
  evalRoll: (dice: number[]) => number;
}

export interface TotalOneNumberProps {
  val: number;
  description: string;
  evalRoll: (dice: number[]) => number;
}

export interface SumDistroProps {
  count: number;
  description: string;
}

export interface RuleInstance extends RuleProps {
  // constructor: any;
  sum: (dice: number[]) => number;
  freq: (dice: number[]) => number[];
  count: (dice: number[], val: number) => number;
}